+++ lib/flutter_addons.dart
@@ -48,19 +48,19 @@ part 'src/context/theme_colors.dart';
 part 'src/context/theme_mode.dart';
 
 // Widgets
-part 'src/widget/widget_ext.dart';
-part 'src/widget/row_column.dart';
-part 'src/widget/stacking.dart';
-part 'src/widget/list_grid.dart';
+part 'src/ui/widget/widget_ext.dart';
+part 'src/ui/widget/row_column.dart';
+part 'src/ui/widget/stacking.dart';
+part 'src/ui/widget/list_grid.dart';
 
 // Responsive
-part 'src/core/responsive/scope.dart';
-part 'src/core/spacing/responsive_space.dart';
-part 'src/core/responsive/screen_utils.dart';
-part 'src/core/responsive/responsive_effects.dart';
+part 'src/core/responsive/entry/entry_scope.dart';
+part 'src/core/responsive/spacing/extended_units.dart';
+part 'src/core/responsive/helpers/screen_utils.dart';
+part 'src/core/responsive/constants/responsive_effects.dart';
 
 // Spacing
-part 'src/core/spacing/spacing_extensions.dart';
+part 'src/core/responsive/spacing/spacing_extensions.dart';
 
 // Routes
 part 'src/routes/navigator_addons.dart';
@@ -99,8 +99,8 @@ part 'src/data/fake_data.dart';
 part 'src/data/img_data.dart';
 
 //  Shapes & Decoration
-part 'src/decoration/decoration.dart';
-part 'src/decoration/button_styles.dart';
+part 'src/core/decoration/decoration.dart';
+part 'src/core/decoration/button_styles.dart';
 
 /// [Soul Theme Engine]
 /// Colors
@@ -134,7 +134,7 @@ part 'src/core/typography/base_styles.dart';
 part 'src/core/typography/fallback_typo.dart';
 
 /// FallBack View
-part 'src/components/ui/error_page.dart';
+part 'src/ui/screens/error_page.dart';
 part 'src/utility/author_speech.dart';
 
 /// Errors
@@ -146,50 +146,62 @@ part 'src/core/errors/retry_mechanism.dart';
 
 /// Components
 /// Alert , Snackbar, Toaster, Dialog, Theme Toggler
-part 'src/components/basic/alart.dart';
-part 'src/components/basic/snackbar.dart';
-part 'src/components/basic/toaster.dart';
-part 'src/components/basic/dialog.dart';
-part 'src/components/widgets/theme_toggler.dart';
-part 'src/decoration/border_box.dart';
+part 'src/ui/extras/alart.dart';
+part 'src/ui/extras/snackbar.dart';
+part 'src/ui/extras/toaster.dart';
+part 'src/ui/extras/dialog.dart';
+part 'src/ui/basic/theme_toggler.dart';
+//part 'src/decoration/border_box.dart';
 
 //responsive helper
-part 'src/core/responsive/design_utils.dart';
+part 'src/core/responsive/helpers/design_utils.dart';
 part 'src/routes/navigator_manager.dart';
 
-part 'src/core/responsive/smart_extensions.dart';
-part 'src/core/responsive/unified_scale.dart';
-part 'src/core/responsive/smartui_utils.dart';
-part 'src/core/responsive/unified_enums.dart';
-part 'src/core/responsive/responsive_builder.dart';
-part 'src/core/responsive/base_scaffold.dart';
+part 'src/core/responsive/helpers/smart_extensions.dart';
+part 'src/core/responsive/entry/core_scale.dart';
+part 'src/core/responsive/helpers/smartui_utils.dart';
+part 'src/core/responsive/constants/responsive_enums.dart';
+part 'src/core/responsive/models/frame_model.dart';
+part 'src/core/responsive/entry/responsive_builder.dart';
+part 'src/core/responsive/entry/base_scaffold.dart';
 // Colors
 part 'src/core/colors/kolors.dart';
 part 'src/core/typography/text_weight.dart';
 part 'src/core/typography/text_size.dart';
-part 'src/components/widgets/label.dart';
-part 'src/components/component/button.dart';
+part 'src/ui/basic/label.dart';
+part 'src/ui/basic/button.dart';
 part 'src/core/colors/safe_colors.dart';
 
 // decoration
-part 'src/decoration/input_decoration.dart';
-part 'src/decoration/shapes.dart';
+part 'src/core/decoration/input_decoration.dart';
+part 'src/core/decoration/shapes.dart';
 
 // Responsive Utilities
-part 'src/core/responsive/summery_logger.dart';
-part 'src/core/responsive/media_info.dart';
-part 'src/core/responsive/orentaion_lock.dart';
+part 'src/core/responsive/helpers/summery_logger.dart';
+part 'src/core/responsive/models/media_info.dart';
+part 'src/core/responsive/models/orentaion_lock.dart';
 part 'src/animation/animate_proxy.dart';
 
+// flexing
+part 'src/core/responsive/bootstrap/breadcrums.dart';
+part 'src/ui/basic/fx_container.dart';
+part 'src/core/responsive/bootstrap/flex_core.dart';
+part 'src/core/responsive/bootstrap/bootstrap_section.dart';
+part 'src/core/responsive/bootstrap/flaxable_spacing.dart';
+
 //components
-part 'src/components/component/appbar_addon.dart';
-part 'src/components/component/avatar_circle.dart';
-part 'src/components/component/page_indicator.dart';
-part 'src/components/ui/mini_progress.dart';
-part 'src/components/shape/box_kit.dart';
+part 'src/ui/basic/appbar_addon.dart';
+part 'src/ui/basic/avatar_circle.dart';
+part 'src/ui/basic/page_indicator.dart';
+part 'src/ui/components/ui_miniprogress.dart';
+part 'src/core/decoration/box_kit.dart';
 
 // state management
 part 'src/core/state/app_error.dart';
 part 'src/core/state/app_state.dart';
 part 'src/core/state/exception.dart';
 part 'src/core/state/functional.dart';
+
+// ui component
+part 'src/ui/components/ui_borderbox.dart';
+part 'src/ui/components/ui_popupmenu.dart';